// src/api/resources/session/client/Client.ts
// This file was auto-generated by Fern from our API Definition.

import type { BaseClientOptions } from "../../../../BaseClient.js";
import * as core from "../../../../core/index.js";
import * as environments from "../../../../environments.js";
import { SessionSocket } from "./Socket.js";

export declare namespace Session {
    export interface Options extends BaseClientOptions {}

    export interface ConnectArgs {
        /** Arbitrary headers to send with the websocket connect request. */
        headers?: Record<string, string>;
        /** Enable debug mode on the websocket. Defaults to false. */
        debug?: boolean;
        /** Number of reconnect attempts. Defaults to 30. */
        reconnectAttempts?: number;
    }
}

export class Session {
    protected readonly _options: Session.Options;

    constructor(_options: Session.Options = {}) {
        this._options = _options;
    }

    public async connect(args: Session.ConnectArgs = {}): Promise<SessionSocket> {
        const { headers, debug, reconnectAttempts } = args;
        const _headers: Record<string, unknown> = { ...headers };
        const socket = new core.ReconnectingWebSocket({
            url: core.url.join(
                (await core.Supplier.get(this._options.baseUrl)) ??
                    ((await core.Supplier.get(this._options.environment)) ?? environments.ApolloEnvironment.Gcp)
                        .production,
                "/messaging/v1/session",
            ),
            protocols: [],
            queryParameters: {},
            headers: _headers,
            options: { debug: debug ?? false, maxRetries: reconnectAttempts ?? 30 },
        });
        return new SessionSocket({ socket });
    }
}